se agregaron las funciones de los botones 1,2,3,4 y se añadio el movimiento de los disparos

+++ cuerpo.cpp
@@ -1,6 +1,6 @@
 #include "cuerpo.h"
 
-cuerpo::cuerpo(int r_, int x, int y)
+cuerpo::cuerpo(float r_, float x, float y)
 {
     r = r_;
     posx = x;
